tweak templates and fix dropdown list missing
