+++ packages/augur-tools/src/templates-source.ts
@@ -2,7 +2,6 @@ import {
   SPORTS,
   GOLF,
   TemplateInputType,
-  LIST_VALUES,
   REQUIRED,
   HOCKEY,
   ValidationType,
@@ -30,6 +29,7 @@ import {
   ETHEREUM,
   LITECOIN,
 } from './templates-template';
+import { LIST_VALUES } from './templates-lists';
 
 const YES_NO = 'YesNo';
 const CATEGORICAL = 'Categorical'
@@ -1494,8 +1494,8 @@ export const TEMPLATES = {
             templates: [
               {
                 marketType: YES_NO,
-                question: `Will the [0] win vs the [1], Estimated schedule start time: [2]?`,
-                example: `Will the NY Giants win vs. the New England Patriots, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NFL: Will the [0] win vs the [1], Estimated schedule start time: [2]?`,
+                example: `NFL:Will the NY Giants win vs. the New England Patriots, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -1529,49 +1529,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: YES_NO,
-                question: `Will the [0] win vs the [1] by [2] or more points, Estimated schedule start time: [3]?`,
-                example: `Will the NY Giants win vs. the New England Patriots by 3 or more points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
-                inputs: [
-                  {
-                    id: 0,
-                    type: TemplateInputType.DROPDOWN,
-                    placeholder: `Team A`,
-                    values: LIST_VALUES.NFL_TEAMS,
-                  },
-                  {
-                    id: 1,
-                    type: TemplateInputType.DROPDOWN,
-                    placeholder: `Team B`,
-                    values: LIST_VALUES.NFL_TEAMS,
-                  },
-                  {
-                    id: 2,
-                    type: TemplateInputType.TEXT,
-                    validationType: ValidationType.WHOLE_NUMBER,
-                    placeholder: `Whole #`,
-                  },
-                  {
-                    id: 3,
-                    type: TemplateInputType.DATETIME,
-                    placeholder: `Date time`,
-                  },
-                ],
-                resolutionRules: {
-                  [REQUIRED]: [
-                    { text: `Include Regulation and Overtime` },
-                    {
-                      text: `If the game ends in a tie, the market should resolve as "NO' as Team A did NOT win vs team B`,
-                    },
-                    {
-                      text: `At least 55 minutes of play must have elapsed for the game to be deemed official.  If less than 55 minutes of play have been completed, there is no official winner of the game and the market should resolve as "No"`,
-                    },
-                  ],
-                },
-              },
-              {
-                marketType: YES_NO,
-                question: `Will the [0] & [1] score [2] or more combined points, Estimated schedule start time: [3]?`,
-                example: `Will the NY Giants & the New England Patriots score 44 or more combined points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NFL: Will the [0] win vs the [1] by [2] or more points, Estimated schedule start time: [3]?`,
+                example: `NFL: Will the NY Giants win vs. the New England Patriots by 3 or more points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -1611,8 +1570,49 @@ export const TEMPLATES = {
               },
               {
                 marketType: YES_NO,
-                question: `Will the [0] have [1] or more regular season wins in [2]?`,
-                example: `Will the Dallas Cowboys have 9 or more regular season wins in 2019?`,
+                question: `NFL: Will the [0] & [1] score [2] or more combined points, Estimated schedule start time: [3]?`,
+                example: `NFL: Will the NY Giants & the New England Patriots score 44 or more combined points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                inputs: [
+                  {
+                    id: 0,
+                    type: TemplateInputType.DROPDOWN,
+                    placeholder: `Team A`,
+                    values: LIST_VALUES.NFL_TEAMS,
+                  },
+                  {
+                    id: 1,
+                    type: TemplateInputType.DROPDOWN,
+                    placeholder: `Team B`,
+                    values: LIST_VALUES.NFL_TEAMS,
+                  },
+                  {
+                    id: 2,
+                    type: TemplateInputType.TEXT,
+                    validationType: ValidationType.WHOLE_NUMBER,
+                    placeholder: `Whole #`,
+                  },
+                  {
+                    id: 3,
+                    type: TemplateInputType.DATETIME,
+                    placeholder: `Date time`,
+                  },
+                ],
+                resolutionRules: {
+                  [REQUIRED]: [
+                    { text: `Include Regulation and Overtime` },
+                    {
+                      text: `If the game ends in a tie, the market should resolve as "NO' as Team A did NOT win vs team B`,
+                    },
+                    {
+                      text: `At least 55 minutes of play must have elapsed for the game to be deemed official.  If less than 55 minutes of play have been completed, there is no official winner of the game and the market should resolve as "No"`,
+                    },
+                  ],
+                },
+              },
+              {
+                marketType: YES_NO,
+                question: `NFL: Will the [0] have [1] or more regular season wins in [2]?`,
+                example: `NFL: Will the Dallas Cowboys have 9 or more regular season wins in 2019?`,
                 inputs: [
                   {
                     id: 0,
@@ -1643,8 +1643,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: YES_NO,
-                question: `Will the [0] win SuperBowl [1]?`,
-                example: `Will the NY Giants win Superbowl LIV?`,
+                question: `NFL: Will the [0] win SuperBowl [1]?`,
+                example: `NFL: Will the NY Giants win Superbowl LIV?`,
                 inputs: [
                   {
                     id: 0,
@@ -1767,8 +1767,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: CATEGORICAL,
-                question: `[0] vs [1]: Total goals scored; Over/Under [2].5, Estimated schedule start time: [3]?`,
-                example: `NY Giants vs Dallas Cowboys: Total points scored: Over/Under 56.5, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NFL: [0] vs [1]: Total goals scored; Over/Under [2].5, Estimated schedule start time: [3]?`,
+                example: `NFL: NY Giants vs Dallas Cowboys: Total points scored: Over/Under 56.5, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -1858,7 +1858,7 @@ export const TEMPLATES = {
                     values: LIST_VALUES.YEARS,
                   },
                   {
-                    id: 0,
+                    id: 1,
                     type: TemplateInputType.DROPDOWN,
                     placeholder: `Award`,
                     values: LIST_VALUES.FOOTBALL_AWARDS,
@@ -1900,8 +1900,8 @@ export const TEMPLATES = {
             templates: [
               {
                 marketType: YES_NO,
-                question: `Will the [0] win vs the [1], Estimated schedule start time: [2]?`,
-                example: `Will Alabama Crimson Tide win vs. Florida Gators, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NCAA FB: Will the [0] win vs the [1], Estimated schedule start time: [2]?`,
+                example: `NCAA FB: Will Alabama Crimson Tide win vs. Florida Gators, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -1933,8 +1933,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: YES_NO,
-                question: `Will the [0] win vs the [1] by [2] or more points, Estimated schedule start time: [3]?`,
-                example: `Will Alabama Crimson Tide win vs. Florida Gators by 7 or more points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NCAA FB: Will the [0] win vs the [1] by [2] or more points, Estimated schedule start time: [3]?`,
+                example: `NCAA FB: Will Alabama Crimson Tide win vs. Florida Gators by 7 or more points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -1972,8 +1972,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: YES_NO,
-                question: `Will [0] & [1] score [2] or more combined points, Estimated schedule start time: [3]?`,
-                example: `Will USC & UCLA score 51 or more combined points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NCAA FB: Will [0] & [1] score [2] or more combined points, Estimated schedule start time: [3]?`,
+                example: `NCAA FB: Will USC & UCLA score 51 or more combined points, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -2049,8 +2049,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: CATEGORICAL,
-                question: `[0] vs [1]: Total points scored; Over/Under [2].5, Estimated schedule start time: [3]?`,
-                example: `Alabama vs Michigan: Total points scored: Over/Under 56.5, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
+                question: `NCAA FB: [0] vs [1]: Total points scored; Over/Under [2].5, Estimated schedule start time: [3]?`,
+                example: `NCAA FB: Alabama vs Michigan: Total points scored: Over/Under 56.5, Estimated schedule start time: Sept 19, 2019 1:00 pm EST?`,
                 inputs: [
                   {
                     id: 0,
@@ -2103,8 +2103,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: CATEGORICAL,
-                question: `Which team will win the [0] [1]: [2] vs [3]?`,
-                example: `Which team will win the 2020 Cotton Bowl: Tennessee Volunteers vs Miami Hurricanes?`,
+                question: `NCAA FB: Which team will win the [0] [1]: [2] vs [3]?`,
+                example: `NCAA FB: Which team will win the 2020 Cotton Bowl: Tennessee Volunteers vs Miami Hurricanes?`,
                 inputs: [
                   {
                     id: 0,
@@ -2115,7 +2115,7 @@ export const TEMPLATES = {
                   {
                     id: 1,
                     type: TemplateInputType.TEXT,
-                    placeholder: `Game`,
+                    placeholder: `Bowl Game`,
                   },
                   {
                     id: 2,
@@ -2195,8 +2195,8 @@ export const TEMPLATES = {
               },
               {
                 marketType: SCALAR,
-                question: `Total number of wins [0] will finish [1] regular season with?`,
-                example: `Total number of wins Michigan Wolverines will finish 2019 regular season with?`,
+                question: `NCAA FB: Total number of wins [0] will finish [1] regular season with?`,
+                example: `NCAA FB: Total number of wins Michigan Wolverines will finish 2019 regular season with?`,
                 denomination: 'wins',
                 tickSize: 1,
                 inputs: [
