+++ packages/augur-tools/src/templates-template.ts
@@ -149,1187 +149,6 @@ export const ValidationTemplateInputType = {
   [TemplateInputType.DATEYEAR]: `(January|February|March|April|May|June|July|August|September|October|November|December) ([0-9]){2}, 20|([0-9]{2})`
 };
 
-export const LIST_VALUES = {
-  YEARS: [
-    {
-      value: '2019',
-      label: '2019',
-    },
-    {
-      value: '2020',
-      label: '2020',
-    },
-    {
-      value: '2021',
-      label: '2021',
-    },
-    {
-      value: '2022',
-      label: '2022',
-    },
-  ],
-  YEAR_RANGE: [
-    {
-      value: '2019-20',
-      label: '2019-20',
-    },
-    {
-      value: '2020-21',
-      label: '2020-21',
-    },
-    {
-      value: '2021-22',
-      label: '2021-22',
-    },
-  ],
-  FOOTBALL_AWARDS: [
-    {
-      value: 'MVP',
-      label: 'MVP',
-    },
-    {
-      value: 'Offensive Player of the year',
-      label: 'Offensive Player of the year',
-    },
-    {
-      value: 'Defensive player of the year',
-      label: 'Defensive player of the year',
-    },
-    {
-      value: 'Offensive Rookie of the year',
-      label: 'Offensive Rookie of the year',
-    },
-    {
-      value: 'Defensive Rookie of the year',
-      label: 'Defensive Rookie of the year',
-    },
-    {
-      value: 'AP Most Valuable Player',
-      label: 'AP Most Valuable Player'
-    }
-  ],
-  FOOTBALL_EVENT: [
-    {
-      value: 'Superbowl',
-      label: 'Superbowl',
-    },
-    {
-      value: 'AFC Championship game',
-      label: 'AFC Championship game',
-    },
-    {
-      value: 'NFC Championship game',
-      label: 'NFC Championship game',
-    },
-  ],
-  NCAA_BASKETBALL_CONF: [
-    {
-      label: 'American East',
-      value: 'American East',
-    },
-    {
-      label: 'American',
-      value: 'American',
-    },
-    {
-      label: 'Atlantic 10',
-      value: 'Atlantic 10',
-    },
-    {
-      label: 'ACC',
-      value: 'ACC',
-    },
-    {
-      label: 'Atlantic Sun',
-      value: 'Atlantic Sun',
-    },
-    {
-      label: 'Big 12',
-      value: 'Big 12',
-    },
-    {
-      label: 'Big East',
-      value: 'Big East',
-    },
-    {
-      label: 'Big Sky',
-      value: 'Big Sky',
-    },
-    {
-      label: 'Big South',
-      value: 'Big South',
-    },
-    {
-      label: 'Big Ten',
-      value: 'Big Ten',
-    },
-    {
-      label: 'Big West',
-      value: 'Big West',
-    },
-    {
-      label: 'Colonial',
-      value: 'Colonial',
-    },
-    {
-      label: 'Conference USA',
-      value: 'Conference USA',
-    },
-    {
-      label: 'Horizon',
-      value: 'Horizon',
-    },
-    {
-      label: 'Ivy',
-      value: 'Ivy',
-    },
-    {
-      label: 'MAAC',
-      value: 'MAAC',
-    },
-    {
-      label: 'Mid-American',
-      value: 'Mid-American',
-    },
-    {
-      label: 'MEAC',
-      value: 'MEAC',
-    },
-    {
-      label: 'Missouri Valley',
-      value: 'Missouri Valley',
-    },
-    {
-      label: 'Mountain West',
-      value: 'Mountain West',
-    },
-    {
-      label: 'Northeast',
-      value: 'Northeast',
-    },
-    {
-      label: 'Ohio Valley',
-      value: 'Ohio Valley',
-    },
-    {
-      label: 'Pac-12',
-      value: 'Pac-12',
-    },
-    {
-      label: 'Patriot League',
-      value: 'Patriot League',
-    },
-    {
-      label: 'SEC',
-      value: 'SEC',
-    },
-    {
-      label: 'Southern',
-      value: 'Southern',
-    },
-    {
-      label: 'Southland',
-      value: 'Southland',
-    },
-    {
-      label: 'SWAC',
-      value: 'SWAC',
-    },
-    {
-      label: 'Summit League',
-      value: 'Summit League',
-    },
-    {
-      label: 'Sun Belt',
-      value: 'Sun Belt',
-    },
-    {
-      label: 'West Coast',
-      value: 'West Coast',
-    },
-    {
-      label: 'WAC',
-      value: 'WAC',
-    },
-  ],
-  BASEBALL_EVENT: [
-    {
-      value: 'American League Division Series',
-      label: 'American League Division Series',
-    },
-    {
-      value: 'National League Division Series',
-      label: 'National League Division Series',
-    },
-    {
-      value: 'American League Championship Series',
-      label: 'American League Championship Series',
-    },
-    {
-      value: 'National League Championship Series',
-      label: 'National League Championship Series',
-    },
-    {
-      value: 'World Series',
-      label: 'World Series',
-    },
-  ],
-  ENTERTAINMENT_EVENT: [
-    {
-      value: 'Academy Awards',
-      label: 'Academy Awards',
-    },
-    {
-      value: 'Emmy Awards',
-      label: 'Emmy Awards',
-    },
-    {
-      value: 'Grammy Awards',
-      label: 'Grammy Awards',
-    },
-    {
-      value: 'Golden Globe Awards',
-      label: 'Golden Globe Awards',
-    },
-  ],
-  CURRENCY: [
-    {
-      value: 'US dollar (USD)',
-      label: 'US dollar (USD)',
-    },
-    {
-      value: 'Euro (EUR)',
-      label: 'Euro (EUR)',
-    },
-    {
-      value: 'Chinese yuan (CNY)',
-      label: 'Chinese yuan (CNY)',
-    },
-    {
-      value: 'British pound (GBP)',
-      label: 'British pound (GBP)',
-    },
-    {
-      value: 'Australian dollar (AUD)',
-      label: 'Australian dollar (AUD)',
-    },
-    {
-      value: 'Canadian dollar (CAD)',
-      label: 'Canadian dollar (CAD)',
-    },
-    {
-      value: 'Swiss franc (CHF)',
-      label: 'Swiss franc (CHF)',
-    },
-  ],
-  REGION: [
-    {
-      value: 'US (United States)',
-      label: 'US (United States)',
-    },
-    {
-      value: 'Worldwide',
-      label: 'Worldwide',
-    },
-  ],
-  POL_PARTY: [
-    {
-      value: 'Democratic',
-      label: 'Democratic',
-    },
-    {
-      value: 'Republican',
-      label: 'Republican',
-    },
-  ],
-  PRES_OFFICES: [
-    {
-      value: 'President',
-      label: 'President',
-    },
-    {
-      value: 'Vice-President',
-      label: 'Vice-President',
-    },
-  ],
-  OFFICES: [
-    {
-      value: 'President',
-      label: 'President',
-    },
-    {
-      value: 'Vice-President',
-      label: 'Vice-President',
-    },
-    {
-      value: 'Senator',
-      label: 'Senator',
-    },
-    {
-      value: 'Congress',
-      label: 'Congress',
-    },
-  ],
-  POL_POSITION: [
-    {
-      value: 'President',
-      label: 'President',
-    },
-    {
-      value: 'Prime Minister',
-      label: 'Prime Minister',
-    },
-    {
-      value: 'Supreme Leader',
-      label: 'Supreme Leader',
-    },
-    {
-      value: 'Crown Prince',
-      label: 'Crown Prince',
-    },
-    {
-      value: 'Chancellor',
-      label: 'Chancellor',
-    },
-    {
-      value: 'Chief minister',
-      label: 'Chief minister',
-    },
-  ],
-  US_STATES: [
-    {
-      value: 'Alabama',
-      label: 'Alabama',
-    },
-    {
-      value: 'Alaska',
-      label: 'Alaska',
-    },
-    {
-      value: 'Arizona',
-      label: 'Arizona',
-    },
-    {
-      value: 'Arkansas',
-      label: 'Arkansas',
-    },
-    {
-      value: 'California',
-      label: 'California',
-    },
-    {
-      value: 'Colorado',
-      label: 'Colorado',
-    },
-    {
-      value: 'Connecticut',
-      label: 'Connecticut',
-    },
-    {
-      value: 'Delaware',
-      label: 'Delaware',
-    },
-    {
-      value: 'Florida',
-      label: 'Florida',
-    },
-    {
-      value: 'Georgia',
-      label: 'Georgia',
-    },
-    {
-      value: 'Hawaii',
-      label: 'Hawaii',
-    },
-    {
-      value: 'Idaho',
-      label: 'Idaho',
-    },
-    {
-      value: 'Illinois',
-      label: 'Illinois',
-    },
-    {
-      value: 'Indiana',
-      label: 'Indiana',
-    },
-    {
-      value: 'Iowa',
-      label: 'Iowa',
-    },
-    {
-      value: 'Kansas',
-      label: 'Kansas',
-    },
-    {
-      value: 'Kentucky',
-      label: 'Kentucky',
-    },
-    {
-      value: 'Louisiana',
-      label: 'Louisiana',
-    },
-    {
-      value: 'Maine',
-      label: 'Maine',
-    },
-    {
-      value: 'Maryland',
-      label: 'Maryland',
-    },
-    {
-      value: 'Massachusetts',
-      label: 'Massachusetts',
-    },
-    {
-      value: 'Michigan',
-      label: 'Michigan',
-    },
-    {
-      value: 'Minnesota',
-      label: 'Minnesota',
-    },
-    {
-      value: 'Mississippi',
-      label: 'Mississippi',
-    },
-    {
-      value: 'Missouri',
-      label: 'Missouri',
-    },
-    {
-      value: 'Montana',
-      label: 'Montana',
-    },
-    {
-      value: 'Nebraska',
-      label: 'Nebraska',
-    },
-    {
-      value: 'Nevada',
-      label: 'Nevada',
-    },
-    {
-      value: 'New Hampshire',
-      label: 'New Hampshire',
-    },
-    {
-      value: 'New Jersey',
-      label: 'New Jersey',
-    },
-    {
-      value: 'New Mexico',
-      label: 'New Mexico',
-    },
-    {
-      value: 'New York',
-      label: 'New York',
-    },
-    {
-      value: 'North Carolina',
-      label: 'North Carolina',
-    },
-    {
-      value: 'North Dakota',
-      label: 'North Dakota',
-    },
-    {
-      value: 'Ohio',
-      label: 'Ohio',
-    },
-    {
-      value: 'Oklahoma',
-      label: 'Oklahoma',
-    },
-    {
-      value: 'Oregon',
-      label: 'Oregon',
-    },
-    {
-      value: 'Pennsylvania',
-      label: 'Pennsylvania',
-    },
-    {
-      value: 'Rhode Island',
-      label: 'Rhode Island',
-    },
-    {
-      value: 'South Carolina',
-      label: 'South Carolina',
-    },
-    {
-      value: 'South Dakota',
-      label: 'South Dakota',
-    },
-    {
-      value: 'Tennessee',
-      label: 'Tennessee',
-    },
-    {
-      value: 'Texas',
-      label: 'Texas',
-    },
-    {
-      value: 'Utah',
-      label: 'Utah',
-    },
-    {
-      value: 'Vermont',
-      label: 'Vermont',
-    },
-    {
-      value: 'Virginia',
-      label: 'Virginia',
-    },
-    {
-      value: 'Washington',
-      label: 'Washington',
-    },
-    {
-      value: 'West Virginia',
-      label: 'West Virginia',
-    },
-    {
-      value: 'Wisconsin',
-      label: 'Wisconsin',
-    },
-    {
-      value: 'Wyoming',
-      label: 'Wyoming',
-    },
-  ],
-  BASKETBALL_EVENT: [
-    {
-      value: 'Eastern Conference Finals',
-      label: 'Eastern Conference Finals',
-    },
-    {
-      value: 'Western Conference Finals',
-      label: 'Western Conference Finals',
-    },
-    {
-      value: 'NBA Championship',
-      label: 'NBA Championship',
-    },
-  ],
-  NBA_BASKETBALL_AWARD: [
-    {
-      value: 'Most Valuable Player',
-      label: 'Most Valuable Player',
-    },
-    {
-      value: 'Rookie of the year',
-      label: 'Rookie of the year',
-    },
-    {
-      value: '6th Man',
-      label: '6th Man',
-    },
-    {
-      value: 'Defensive Player of the Year',
-      label: 'Defensive Player of the Year',
-    },
-    {
-      value: 'Most Improved player',
-      label: 'Most Improved player',
-    },
-  ],
-  NCAA_BASKETBALL_EVENTS: [
-    {
-      value: 'NCAA Tournament',
-      label: 'NCAA Tournament',
-    },
-    {
-      value: 'Sweet 16',
-      label: 'Sweet 16',
-    },
-    {
-      value: 'Elite 8',
-      label: 'Elite 8',
-    },
-    {
-      value: 'Final Four',
-      label: 'Final Four',
-    },
-  ],
-  NCAA_BASKETBALL_AWARD: [
-    {
-      value: 'Most Valuable Player',
-      label: 'Most Valuable Player',
-    },
-    {
-      value: 'Defensive Player of the Year',
-      label: 'Defensive Player of the Year',
-    },
-    {
-      value: 'Most Improved player',
-      label: 'Most Improved player',
-    },
-    {
-      value: 'Coach of the year',
-      label: 'Coach of the year',
-    },
-    {
-      value: 'NCAA-Naismith College Player of the Year',
-      label: 'NCAA-Naismith College Player of the Year',
-    },
-    {
-      value: 'AP college Player of the year',
-      label: 'AP college Player of the year',
-    },
-    {
-      value: 'Sporting News College Basketball Player of the year',
-      label: 'Sporting News College Basketball Player of the year',
-    },
-    {
-      value: 'John Wooden',
-      label: 'John Wooden',
-    },
-    {
-      value: 'Oscar Robertson',
-      label: 'Oscar Robertson',
-    },
-  ],
-  MENS_WOMENS: [
-    {
-      value: "Men's",
-      label: "Men's",
-    },
-    {
-      value: "Women's",
-      label: "Women's",
-    },
-  ],
-  BASKETBALL_ACTION: [
-    {
-      value: 'Points Scored',
-      label: 'Points Scored',
-    },
-    {
-      value: 'Rebounds',
-      label: 'Rebounds',
-    },
-    {
-      value: 'Assists',
-      label: 'Assists',
-    },
-    {
-      value: 'made 3-pointers',
-      label: 'made 3-pointers',
-    },
-  ],
-  HOCKEY_AWARD: [
-    {
-      value: 'Hart Trophy',
-      label: 'Hart Trophy',
-    },
-    {
-      value: 'Norris Trophy',
-      label: 'Norris Trophy',
-    },
-    {
-      value: 'Vezina Trophy',
-      label: 'Vezina Trophy',
-    },
-    {
-      value: 'Calder Trophy',
-      label: 'Calder Trophy',
-    },
-  ],
-  HORSE_RACING_EVENT: [
-    {
-      value: 'Kentucky Derby',
-      label: 'Kentucky Derby',
-    },
-    {
-      value: 'Preakness',
-      label: 'Preakness',
-    },
-    {
-      value: 'Belmont',
-      label: 'Belmont',
-    },
-    {
-      value: 'Triple Crown',
-      label: 'Triple Crown',
-    },
-  ],
-  TENNIS_EVENT: [
-    {
-      value: 'Australian Open',
-      label: 'Australian Open',
-    },
-    {
-      value: 'French Open',
-      label: 'French Open',
-    },
-    {
-      value: 'Wimbledon',
-      label: 'Wimbledon',
-    },
-    {
-      value: 'US Open',
-      label: 'US Open',
-    },
-  ],
-  GOLF_EVENT: [
-    {
-      value: 'PGA Championship',
-      label: 'PGA Championship',
-    },
-    {
-      value: 'Open Championship',
-      label: 'Open Championship',
-    },
-    {
-      value: 'US Open',
-      label: 'US Open',
-    },
-    {
-      value: 'Masters Tournament',
-      label: 'Masters Tournament',
-    },
-    {
-      value: 'British Open',
-      label: 'British Open',
-    },
-  ],
-  MLB_TEAMS: [
-    {
-      value: 'Arizona Diamondbacks',
-      label: 'Arizona Diamondbacks',
-    },
-    {
-      value: 'Atlanta Braves',
-      label: 'Atlanta Braves',
-    },
-    {
-      value: 'Baltimore Orioles',
-      label: 'Baltimore Orioles',
-    },
-    {
-      value: 'Boston Red Sox',
-      label: 'Boston Red Sox',
-    },
-    {
-      value: 'Chicago White Sox',
-      label: 'Chicago White Sox',
-    },
-    {
-      value: 'Chicago Cubs',
-      label: 'Chicago Cubs',
-    },
-    {
-      value: 'Cincinnati Reds',
-      label: 'Cincinnati Reds',
-    },
-    {
-      value: 'Cleveland Indians',
-      label: 'Cleveland Indians',
-    },
-    {
-      value: 'Colorado Rockies',
-      label: 'Colorado Rockies',
-    },
-    {
-      value: 'Detroit Tigers',
-      label: 'Detroit Tigers',
-    },
-    {
-      value: 'Houston Astros',
-      label: 'Houston Astros',
-    },
-    {
-      value: 'Kansas City Royals',
-      label: 'Kansas City Royals',
-    },
-    {
-      value: 'Las Angeles Angels',
-      label: 'Las Angeles Angels',
-    },
-    {
-      value: 'Las Angeles Dodgers',
-      label: 'Las Angeles Dodgers',
-    },
-    {
-      value: 'Miami Marlins',
-      label: 'Miami Marlins',
-    },
-    {
-      value: 'Milwaukee Brewers',
-      label: 'Milwaukee Brewers',
-    },
-    {
-      value: 'Minnesota Twins',
-      label: 'Minnesota Twins',
-    },
-    {
-      value: 'New York Yankees',
-      label: 'New York Yankees',
-    },
-    {
-      value: 'New York Mets',
-      label: 'New York Mets',
-    },
-    {
-      value: 'Oakland Athletics',
-      label: 'Oakland Athletics',
-    },
-    {
-      value: 'Philadelphia Phillies',
-      label: 'Philadelphia Phillies',
-    },
-    {
-      value: 'Pittsburgh Pirates',
-      label: 'Pittsburgh Pirates',
-    },
-    {
-      value: 'San Diego Padres',
-      label: 'San Diego Padres',
-    },
-    {
-      value: 'San Francisco Giants',
-      label: 'San Francisco Giants',
-    },
-    {
-      value: 'Seattle Mariners',
-      label: 'Seattle Mariners',
-    },
-    {
-      value: 'St. Louis Cardinals',
-      label: 'St. Louis Cardinals',
-    },
-    {
-      value: 'Tampa Bay Rays',
-      label: 'Tampa Bay Rays',
-    },
-    {
-      value: 'Texas Rangers',
-      label: 'Texas Rangers',
-    },
-    {
-      value: 'Toronto Blue Jays',
-      label: 'Toronto Blue Jays',
-    },
-    {
-      value: 'Washington Nationals',
-      label: 'Washington Nationals',
-    },
-  ],
-  NFL_TEAMS: [
-    {
-      value: 'Arizona Cardinals',
-      label: 'Arizona Cardinals',
-    },
-    {
-      value: 'Atlanta Falcons',
-      label: 'Atlanta Falcons',
-    },
-    {
-      value: 'Baltimore Ravens',
-      label: 'Baltimore Ravens',
-    },
-    {
-      value: 'Buffalo Bills',
-      label: 'Buffalo Bills',
-    },
-    {
-      value: 'Carolina Panthers',
-      label: 'Carolina Panthers',
-    },
-    {
-      value: 'Chicago Bears',
-      label: 'Chicago Bears',
-    },
-    {
-      value: 'Cincinnati Bengals',
-      label: 'Cincinnati Bengals',
-    },
-    {
-      value: 'Cleveland Browns',
-      label: 'Cleveland Browns',
-    },
-    {
-      value: 'Dallas Cowboys',
-      label: 'Dallas Cowboys',
-    },
-    {
-      value: 'Denver Broncos',
-      label: 'Denver Broncos',
-    },
-    {
-      value: 'Detroit Lions',
-      label: 'Detroit Lions',
-    },
-    {
-      value: 'Green Bay Packers',
-      label: 'Green Bay Packers',
-    },
-    {
-      value: 'Houston Texans',
-      label: 'Houston Texans',
-    },
-    {
-      value: 'Indianapolis Colts',
-      label: 'Indianapolis Colts',
-    },
-    {
-      value: 'Jacksonville Jaguars',
-      label: 'Jacksonville Jaguars',
-    },
-    {
-      value: 'Kansas City Chiefs',
-      label: 'Kansas City Chiefs',
-    },
-    {
-      value: 'Las Angeles Chargers',
-      label: 'Las Angeles Chargers',
-    },
-    {
-      value: 'Las Angeles Rams',
-      label: 'Las Angeles Rams',
-    },
-    {
-      value: 'Miami Dolphins',
-      label: 'Miami Dolphins',
-    },
-    {
-      value: 'Minnesota Vikings',
-      label: 'Minnesota Vikings',
-    },
-    {
-      value: 'New England Patriots',
-      label: 'New England Patriots',
-    },
-    {
-      value: 'New Orleans Saints',
-      label: 'New Orleans Saints',
-    },
-    {
-      value: 'New York Giants',
-      label: 'New York Giants',
-    },
-    {
-      value: 'New York Jets',
-      label: 'New York Jets',
-    },
-    {
-      value: 'Oakland Raiders',
-      label: 'Oakland Raiders',
-    },
-    {
-      value: 'Philadelphia Eagles',
-      label: 'Philadelphia Eagles',
-    },
-    {
-      value: 'Pittsburgh Steelers',
-      label: 'Pittsburgh Steelers',
-    },
-    {
-      value: 'San Francisco 49ers',
-      label: 'San Francisco 49ers',
-    },
-    {
-      value: 'Seattle Seahawks',
-      label: 'Seattle Seahawks',
-    },
-    {
-      value: 'Tampa Bay Buccaneers',
-      label: 'Tampa Bay Buccaneers',
-    },
-    {
-      value: 'Tennessee Titans',
-      label: 'Tennessee Titans',
-    },
-    {
-      value: 'Washington Redskins',
-      label: 'Washington Redskins',
-    },
-  ],
-  NBA_TEAMS: [
-    {
-      value: 'Atlanta Hawks',
-      label: 'Atlanta Hawks',
-    },
-    {
-      value: 'Boston Celtics',
-      label: 'Boston Celtics',
-    },
-    {
-      value: 'Brooklyn Nets',
-      label: 'Brooklyn Nets',
-    },
-    {
-      value: 'Charlotte Hornets',
-      label: 'Charlotte Hornets',
-    },
-    {
-      value: 'Chicago Bulls',
-      label: 'Chicago Bulls',
-    },
-    {
-      value: 'Cleveland Cavaliers',
-      label: 'Cleveland Cavaliers',
-    },
-    {
-      value: 'Dallas Mavericks',
-      label: 'Dallas Mavericks',
-    },
-    {
-      value: 'Denver Nuggets',
-      label: 'Denver Nuggets',
-    },
-    {
-      value: 'Detroit Pistons',
-      label: 'Detroit Pistons',
-    },
-    {
-      value: 'Golden State Warriors',
-      label: 'Golden State Warriors',
-    },
-    {
-      value: 'Houston Rockets',
-      label: 'Houston Rockets',
-    },
-    {
-      value: 'Indiana Pacers',
-      label: 'Indiana Pacers',
-    },
-    {
-      value: 'LA Clippers',
-      label: 'LA Clippers',
-    },
-    {
-      value: 'LA Lakers',
-      label: 'LA Lakers',
-    },
-    {
-      value: 'Memphis Grizzlies',
-      label: 'Memphis Grizzlies',
-    },
-    {
-      value: 'Miami Heat',
-      label: 'Miami Heat',
-    },
-    {
-      value: 'Milwaukee Bucks',
-      label: 'Milwaukee Bucks',
-    },
-    {
-      value: 'Minnesota Timberwolves',
-      label: 'Minnesota Timberwolves',
-    },
-    {
-      value: 'New Orleans Pelicans',
-      label: 'New Orleans Pelicans',
-    },
-    {
-      value: 'New York Knicks',
-      label: 'New York Knicks',
-    },
-    {
-      value: 'Oklahoma City Thunder',
-      label: 'Oklahoma City Thunder',
-    },
-    {
-      value: 'Orlando Magic',
-      label: 'Orlando Magic',
-    },
-    {
-      value: 'Philadelphia 76ers',
-      label: 'Philadelphia 76ers',
-    },
-    {
-      value: 'Phoenix Suns',
-      label: 'Phoenix Suns',
-    },
-    {
-      value: 'Portland Trail Blazers',
-      label: 'Portland Trail Blazers',
-    },
-    {
-      value: 'Sacramento Kings',
-      label: 'Sacramento Kings',
-    },
-    {
-      value: 'San Antonio Spurs',
-      label: 'San Antonio Spurs',
-    },
-    {
-      value: 'Toronto Raptors',
-      label: 'Toronto Raptors',
-    },
-    {
-      value: 'Utah Jazz',
-      label: 'Utah Jazz',
-    },
-    {
-      value: 'Washington Wizards',
-      label: 'Washington Wizards',
-    },
-  ],
-  BTC_USD_EXCHANGES: [
-    {
-      value: 'Coinbase Pro (pro.coinbase.com)',
-      label: 'Coinbase Pro (pro.coinbase.com)',
-    },
-    {
-      value: 'Bitstamp (bitstamp.net)',
-      label: 'Bitstamp (bitstamp.net)',
-    },
-    {
-      value: 'Bittrex (bittrex.com)',
-      label: 'Bittrex (bittrex.com)',
-    },
-  ],
-  NON_BTC_USD_EXCHANGES: [
-    {
-      value: 'Coinbase Pro (pro.coinbase.com)',
-      label: 'Coinbase Pro (pro.coinbase.com)',
-    },
-    {
-      value: 'Bitstamp (bitstamp.net)',
-      label: 'Bitstamp (bitstamp.net)',
-    },
-    {
-      value: 'Kraken (kraken.com)',
-      label: 'Kraken (kraken.com)',
-    },
-    {
-      value: 'Bittrex (bittrex.com)',
-      label: 'Bittrex (bittrex.com)',
-    },
-  ],
-  USDT_EXCHANGES: [
-    {
-      value: 'Bittrex (bittrex.com)',
-      label: 'Bittrex (bittrex.com)',
-    },
-    {
-      value: 'Binance (binance.com)',
-      label: 'Binance (binance.com)',
-    },
-    {
-      value: 'Huobi Global (hbg.com)',
-      label: 'Huobi Global (hbg.com)',
-    },
-  ],
-  EUR_EXCHANGES: [
-    {
-      value: 'Coinbase Pro (pro.coinbase.com)',
-      label: 'Coinbase Pro (pro.coinbase.com)',
-    },
-    {
-      value: 'Bitstamp (bitstamp.net)',
-      label: 'Bitstamp (bitstamp.net)',
-    },
-    {
-      value: 'Kraken (kraken.com)',
-      label: 'Kraken (kraken.com)',
-    },
-  ],
-};
-
 
 //##TEMPLATES##
